requirements.txt
python-socketio>=5.5.0,<6.0.0
requests>=2.25.1,<3.0.0
rpi-ws281x>=4.3.1,<5.0.0
python-dotenv>=0.19.2
websocket-client>=1.2.3,<2.0.0